+++ packages/las-sdk-core/src/types/log.ts
@@ -1,0 +1,9 @@
+import { RequestConfig } from './common';
+
+export type Log = {
+  logId: string;
+  events: Array<Record<any, any>>;
+  transitionId?: string | null;
+};
+
+export type GetLogOptions = RequestConfig;